refactor(ShadowServer): Move Connector class to own file
